// src/api/generated/model/contestMode.ts
/**
 * Generated by orval v8.8.0 🍺
 * Do not edit manually.
 * Amrita ICPC Coding Platform - Backend
 * A backend tool for Amrita ICPC Coding Platform
 * OpenAPI spec version: 1.0.0
 */

/**
 * Enumeration of team mode for a contest

Attributes:
    INDIVIDUAL: Each participant competes independently.
    TEAM: Teams compete collaboratively.
 */
export type ContestMode = (typeof ContestMode)[keyof typeof ContestMode];

export const ContestMode = {
    individual: "individual",
    team: "team",
} as const;